# openapi/consumer-secrets.yaml
openapi: 3.1.0
info:
  title: Hemlig Secrets Delivery API
  version: 0.4.0
  description: |
    Lambda/S3/KMS backed delivery of externally managed secret payloads.
    This contract includes secret control, organizational browsing, and enrolled
    identity certificate lifecycle routes. Audit query and issuer-root rotation are
    intentionally absent.
...
servers:
  - &admin-server
    url: https://{adminFqdn}
    description: Administrator API; JWT required for every operation.
    variables:
      adminFqdn:
        default: admin.example.com
  - &consumer-server
    url: https://{apiFqdn}
    description: Delivery API; mTLS identity required for every operation.
    variables:
      apiFqdn:
        default: api.example.com
...
paths:
...
  /v1/changes:
    get:
      servers: [*consumer-server]
      tags: [consumer]
      summary: Return the caller consumer's current paginated access snapshot
      security: [{ consumerMtls: [] }]
      parameters:
        - name: cursor
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Current change page
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ChangesResponse" }
        default:
          $ref: "#/components/responses/Error"
  /v1/agent/config:
    get:
      servers: [*consumer-server]
      tags: [agent]
      summary: Return the active AgentGrant and exact private AWS IoT topic
      security: [{ consumerMtls: [] }]
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AgentConfig" }
        default: { $ref: "#/components/responses/Error" }
  /v1/agent/secrets:
    post:
      servers: [*consumer-server]
      tags: [agent]
      summary: Create a write-scoped empty secret with the caller's initial read ACL
      security: [{ consumerMtls: [] }]
      parameters: [{ $ref: "#/components/parameters/IdempotencyKey" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateAgentSecretRequest" }
      responses:
        "201": { $ref: "#/components/responses/ControlRevision" }
        default: { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    administratorJwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
    consumerMtls:
      type: mutualTLS
    bootstrapCapability:
      type: apiKey
      in: header
      name: Authorization
      description: Exact `Bootstrap hmlb_…` one-use capability; never an administrator JWT.
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 8, maxLength: 128 }
  responses:
    ControlRevision:
      description: Current control revision
      headers:
        ETag:
          schema: { type: string }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ControlRevision" }
...
    Error:
      description: Safe application error envelope
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
  schemas:
...
    Metadata:
      type: object
      properties:
        description: { type: string, maxLength: 1024 }
        path:
          type: string
          maxLength: 256
          pattern: "^[a-z0-9][a-z0-9._-]{0,63}(/[a-z0-9][a-z0-9._-]{0,63})*$"
        tags:
          type: object
          maxProperties: 20
          additionalProperties:
            type: string
            maxLength: 128
    Grant:
      type: object
      required: [consumerId, permissions]
      properties:
        consumerId: { type: string }
        permissions:
          type: array
          items: { type: string, enum: [read] }
...
    ControlRevision:
      type: object
      required:
        [
          schemaVersion,
          secretId,
          controlVersionId,
          environment,
          state,
          createdAt,
          createdBy,
          metadata,
          acl,
        ]
      properties:
        schemaVersion: { type: integer, enum: [1] }
        secretId: { type: string }
        controlVersionId: { type: string }
        payloadVersionId: { type: string }
        payloadKeyCount:
          {
            type: integer,
            minimum: 0,
            description: Number of payload entries; plaintext names are not retained.,
          }
        environment: { type: string }
        state: { type: string, enum: [PENDING_VALUE, ACTIVE, REVOKED] }
        createdAt: { type: string, format: date-time }
        createdBy: { $ref: "#/components/schemas/Actor" }
        metadata: { $ref: "#/components/schemas/Metadata" }
        acl: { type: array, items: { $ref: "#/components/schemas/Grant" } }
    Actor:
      type: object
      required: [type, id]
      properties:
        type: { type: string, enum: [human, consumer, system] }
        id: { type: string }
        tenantId: { type: string }
        consumerId: { type: string }
        environment: { type: string }
...
    Change:
      type: object
      required: [secretId, controlVersionId, state, changeKind]
      properties:
        secretId: { type: string }
        controlVersionId: { type: string }
        payloadVersionId: { type: string }
        state: { type: string, enum: [PENDING_VALUE, ACTIVE, REVOKED] }
        changeKind: { type: string, enum: [secret.changed, secret.revoked] }
    ChangesResponse:
      type: object
      required: [changes, generatedAt]
      properties:
        changes: { type: array, items: { $ref: "#/components/schemas/Change" } }
        nextCursor: { type: string }
        generatedAt: { type: string, format: date-time }
...
    AgentConfig:
      type: object
      required: [consumerId, environment, grant, mqtt]
      properties:
        consumerId: { type: string }
        environment: { type: string }
        grant:
          type: object
          required: [grantId, capabilities, readPathPrefixes, writePathPrefixes]
          properties:
            grantId: { type: string }
            capabilities: { type: array, items: { type: string, enum: [read, write] } }
            readPathPrefixes: { type: array, items: { type: string } }
            writePathPrefixes: { type: array, items: { type: string } }
        mqtt:
          type: object
          required: [endpoint, clientId, topic]
          properties:
            endpoint: { type: string }
            clientId: { type: string }
            topic: { type: string }
    CreateAgentSecretRequest:
      type: object
      required: [secretId, metadata]
      properties:
        secretId: { type: string, pattern: "^[a-z][a-z0-9-]{2,63}$" }
        metadata: { $ref: "#/components/schemas/Metadata" }
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, correlationId]
          properties:
            code:
              type: string
              description: |
                Stable machine-readable failure code. `enrollment_failed` is a
                terminal truststore publication failure and must not be
                retried; `conflict` on an administrator mutation means the
                idempotency key was already used and the outcome must be
                reconciled by re-reading the resource.
              enum:
                [
                  bad_request,
                  forbidden,
                  not_found,
                  conflict,
                  enrollment_failed,
                  precondition_failed,
                  internal_error,
                  service_unavailable,
                ]
            message: { type: string }
            correlationId: { type: string }